Removed old physics. Made Player Movement and added in imported new sprites into my file.
Removed the old code for the physics of my character.

Changed the PlayerMovement script to take Horizontal input to control left and right and added in a float variable to control player speed so I can edit it in Unity.
Also added in a if statement to check space inputs so the player can jump.