Created a saw trap. Created a health collectible that heals the player.
Created a saw trap which moves and damages the player if they collide with it.
Created an animation for the saw.
Sprites used for the animation can be found at https://assetstore.unity.com/packages/2d/characters/pixel-adventure-1-155360
Pixel Adventure 1 by Pixel Frog

Created a health collectible that the player can pick up to heal themselves.