Added invincibility and red flash for my character after they take damage. Created prefabs for Collectibles and Traps.
Added invincibility for my character after they take damage. Added in a red flash for when they take damage to clearly indicate to the player that they have taken damage.

Created a prefab for my Collectibles and Traps

Moved bullets to foreground to fix issue when they are not visible.