Added in way for player to change size. Moved pngs from resources folder to tilemap folder.
Added in a mechanic which lets my player shrink. Added in a currentsize float to do calculations on PlayerMovement.
Click shift to let the player change their current size.

Added in a if statement in PlayerAttack script to check if players size is 3 which means they will not be able to shoot when small.

Moved pngs from the resources folder to the tilemap folder and renamed them so tilemap items did not have the same name in the tilemap folder.