Added in weighted jump and coyote time. Commented out dash code
Added in weighted jump so player jumps depending on how long they hold down the jump input.
Added in coyote time so the player has time to jump after they jump off the edge.
Added in code for dash but it is currently not working so I commented it out.